requirements.txt
argparse==1.4.0
paho_mqtt==1.6.1
pyinstaller
joblib==1.2.0
pyaml==23.5.9
scikit-learn==1.2.2